docs + cleanup
