+++ src/pkg/packager/create.go
@@ -7,25 +7,10 @@ package packager
 import (
 	"fmt"
 	"os"
-	"path/filepath"
 
 	"github.com/defenseunicorns/zarf/src/config"
 	"github.com/defenseunicorns/zarf/src/internal/packager/validate"
-	"github.com/defenseunicorns/zarf/src/pkg/message"
 )
-
-func (p *Packager) cdToBaseDir(base string, cwd string) error {
-	if err := os.Chdir(base); err != nil {
-		return fmt.Errorf("unable to access directory %q: %w", base, err)
-	}
-	message.Note(fmt.Sprintf("Using build directory %s", base))
-
-	// differentials are relative to the current working directory
-	if p.cfg.CreateOpts.DifferentialData.DifferentialPackagePath != "" {
-		p.cfg.CreateOpts.DifferentialData.DifferentialPackagePath = filepath.Join(cwd, p.cfg.CreateOpts.DifferentialData.DifferentialPackagePath)
-	}
-	return nil
-}
 
 // Create generates a Zarf package tarball for a given PackageConfig and optional base directory.
 func (p *Packager) Create() (err error) {
